refactor(web): move command palette from pages/_shared to components
Relocate the command-palette module to the shared components directory and
update all importers, with no behavior change. This is the first step of
dismantling pages/_shared so generic, app-wide UI lives under components.
