+++ web/src/components/command-palette/CommandPalette.tsx
