+++ web/src/components/command-palette/CommandPaletteTrigger.tsx
