+++ web/src/components/command-palette/PaletteContext.tsx
