+++ web/src/components/command-palette/command-palette.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/variables' as *;
-@use '../../../styles/mixins' as *;
+@use '../../styles/variables' as *;
+@use '../../styles/mixins' as *;
 
 // ─── Command palette trigger pill ────────────────────────────────────────────
