+++ web/src/components/command-palette/fuzzy.test.ts
