+++ web/src/components/command-palette/usePaletteShortcut.ts
