+++ web/src/pages/builtin/inspect/InspectPage.tsx
@@ -3,8 +3,8 @@ import { toaster } from '../../../utils';
 import { API } from '../../../api';
 import type { InstanceInfo } from '../../../api/inspect';
 import { PageLayout, PageLoader, EmptyState, CommandPaletteHeader } from '../../../components';
-import { usePalette } from '../../_shared/command-palette';
-import type { Command } from '../../_shared/command-palette';
+import { usePalette } from '../../../components/command-palette';
+import type { Command } from '../../../components/command-palette';
 import { InspectPageFooter } from './InspectPageFooter';
 import { InstanceCard } from './InstanceCard';
 import './inspect.scss';
