+++ web/src/pages/modules/pdump/PdumpPage.tsx
@@ -17,8 +17,8 @@ import ConfigStrip from './ConfigStrip';
 import PacketDrawer from './PacketDrawer';
 import DeleteConfigDialog from './DeleteConfigDialog';
 import type { PdumpConfigInfo, CapturedPacket } from './types';
-import { usePalette } from '../../_shared/command-palette';
-import type { Command } from '../../_shared/command-palette';
+import { usePalette } from '../../../components/command-palette';
+import type { Command } from '../../../components/command-palette';
 import { useTabCycle } from '../../_shared/useTabCycle';
 import '../../../styles/draft-page.scss';
 import './pdump.scss';
