+++ web/src/pages/operators/neighbours/NeighboursPage.tsx
@@ -5,8 +5,8 @@ import { Button, Icon } from '@gravity-ui/uikit';
 import { Plus, Layers } from '@gravity-ui/icons';
 import { PageLayout, PageLoader, ConfigTabStrip, BulkBar, EmptyPagePlaceholder } from '../../../components';
 import { BulkDeleteModal, DeleteConfigModal, CommandPaletteHeader } from '../../../components';
-import { usePalette } from '../../_shared/command-palette';
-import type { Command, RowAdapter } from '../../_shared/command-palette';
+import { usePalette } from '../../../components/command-palette';
+import type { Command, RowAdapter } from '../../../components/command-palette';
 import { useTabCycle } from '../../_shared/useTabCycle';
 import { stringToIPAddress, ipAddressToString } from '../../../utils/netip';
 import { parseIPAddress } from '../../../utils';
@@ -21,7 +21,7 @@ import type { SortState, SortableColumn } from './types';
 import { FamilyFilter, type IPFamily } from '../../../components/VirtualTable';
 import { nudStateToName, STATE_META } from './stateMeta';
 import '../../../styles/draft-page.scss';
-import '../../_shared/command-palette/command-palette.scss';
+import '../../../components/command-palette/command-palette.scss';
 
 const QP_TAB = 'tab';
 const QP_SORT = 'sort';
